Port to 1.20 ItemRendererMixin -> DrawContextMixin
