+++ src/main/resources/simpleshulkerpreview.mixins.json
@@ -6,7 +6,7 @@
   "mixins": [
   ],
   "client": [
-    "ItemRendererMixin"
+    "DrawContextMixin"
   ],
   "injectors": {
     "defaultRequire": 1
